// sentry/src/app/types/type.ts
export type Theme = 'light' | 'dark';
export type Mode = 'min' | 'max';
export type StatCardColourScheme = 'sky' | 'teal' | 'indigo' | 'fuchsia';
export type ChartDataColour =
    | 'sky'
    | 'fuchsia'
    | 'teal'
    | 'rose'
    | 'indigo'
    | 'orange'
    | 'violet';
export type Booleany = 'falsey' | 'truthy';